feat: separate implementation from headers and add thread count inspection
